requirements.txt
django>=4.0,<5.0
gunicorn>=20.0,<21.0
dj-database-url>=1.0,<2.0
whitenoise>=6.0,<7.0
psycopg2>=2.0,<3.0
asgiref>=3.2.7
django-model-utils==4.0.0
pytz==2019.3
sqlparse==0.3.1
Unipath==1.1
djangorestframework~=3.13.1